Lag egen service for generering av brev for forenklet tilbakekrevingsvedtak for å unngå circular dependency
